+++ .gitignore
@@ -16,3 +16,6 @@ node_modules/
 
 # dotenv environment variables file
 .env
+
+# config.json file
+config.json
