+++ package.json
@@ -4,12 +4,15 @@
   "description": "BINGOs people. Gives them the bingo role and displays a message with their pfp being crushed by a tombstone. B I N G O. Discord.JS bot",
   "main": "index.js",
   "scripts": {
-    "start": "node index.js"
+    "start": "ts-node index.ts"
   },
   "author": "SeeBeyond",
   "license": "MIT",
   "dependencies": {
     "discord.js": "^11.4.2",
     "gifencoder": "^2.0.1"
+  },
+  "devDependencies": {
+    "@types/node": "^10.12.24"
   }
 }
